test([issue-2918]): extract shared sprite candidate/lock fixtures
reference.test.js, walk.test.js, and atlas.test.js each hand-rolled the same
candidate-PNG + placeCandidate + lock-anchors chain. A change to the
lock/candidate flow needed three parallel edits. Extract the chain into
spriteTestFixtures.js, parameterized on each suite's TEST_ROOT.

walkTrims.test.js does not actually use this fixture chain (it builds its
own strip-frame fixture for a different purpose), so it is untouched.
