+++ server/services/sprites/walk.test.js
@@ -10,9 +10,9 @@ import { describe, it, expect, vi, beforeEach, afterAll } from 'vitest';
 import { mkdtempSync, rmSync } from 'fs';
 import { tmpdir } from 'os';
 import { join } from 'path';
-import sharp from 'sharp';
 import { mkdir, writeFile, readFile } from 'fs/promises';
 import { createHash } from 'crypto';
+import { lockAllAnchors } from './spriteTestFixtures.js';
 
 const TEST_ROOT = mkdtempSync(join(tmpdir(), 'sprite-walk-test-'));
 
@@ -77,30 +77,9 @@ let seq = 0;
 const newId = () => `walker-${++seq}`;
 const sha256 = (buf) => createHash('sha256').update(buf).digest('hex');
 
-async function writeCandidatePng(path) {
-  const w = 64; const h = 64;
-  const buf = Buffer.alloc(w * h * 3);
-  for (let p = 0; p < w * h; p++) buf.set([255, 0, 255], p * 3);
-  for (let y = 10; y < 40; y++) for (let x = 20; x < 30; x++) buf.set([23, 107, 101], (y * w + x) * 3);
-  await mkdir(join(path, '..'), { recursive: true });
-  await sharp(buf, { raw: { width: w, height: h, channels: 3 } }).png().toFile(path);
-}
-
-async function placeCandidate(recordId, target, name) {
-  const candDir = join(TEST_ROOT, 'sprites', recordId, 'reference', 'candidates');
-  await writeCandidatePng(join(candDir, name));
-  await writeFile(join(candDir, `${name.replace(/\.png$/, '')}.generation.json`), JSON.stringify({
-    schemaVersion: 1, target, chromaKey: '#FF00FF',
-  }));
-  return `reference/candidates/${name}`;
-}
-
 async function characterWithLockedAnchors(id, directions = ['east']) {
   await records.createRecord({ kind: 'character', name: 'Walker' }, id);
-  await lockReference(id, { target: 'main', candidate: await placeCandidate(id, 'main', 'walk-south-candidate-01.png') });
-  for (const dir of directions.filter((d) => d !== 'south')) {
-    await lockReference(id, { target: dir, candidate: await placeCandidate(id, dir, `walk-${dir}-candidate-01.png`) });
-  }
+  await lockAllAnchors(TEST_ROOT, id, { lockReference, directions });
   return id;
 }
